First stab at a refactor
